FST 83 some refactoring structuring
