+++ src/features/integration/IntegrationConfigurationForm.tsx
@@ -4,12 +4,12 @@ import {RouteComponentProps, withRouter} from "react-router-dom";
 import {Accordion, AccordionDetails, AccordionSummary, Box, Theme, Typography} from "@mui/material";
 import {createStyles, makeStyles} from "@mui/styles";
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import RecordForm from "./components/RecordForm";
-import DocumentForm from "./components/DocumentForm";
-import ApplicantForm from "./components/ApplicantForm";
-import CaseForm from "./components/CaseForm";
+import RecordForm from "./components/form/RecordForm";
+import DocumentForm from "./components/form/DocumentForm";
+import ApplicantForm from "./components/form/ApplicantForm";
+import CaseForm from "./components/form/CaseForm";
 import IFormData from "./types/Form/FormData";
-import CaseInformation from "./components/CaseInformation";
+import CaseInformation from "./components/form/CaseInformation";
 import IntegrationService from "./service/IntegrationService";
 import {defaultValues} from "./util/DefaultValues";
 import {mapToDto} from "./util/MapToDto";
