+++ src/features/integration/types/Field.ts
@@ -1,10 +1,11 @@
-import {property, VALUE_BUILDER_STRATEGY} from "../util/util";
+import {VALUE_BUILDER_STRATEGY} from "./ValueBuilderStrategy.enum";
+import {Property} from "./Property";
 
 export interface IField {
     field?: string;
     valueBuildStrategy?: VALUE_BUILDER_STRATEGY,
     valueBuilder?: {
         value?: string;
-        properties?: property[]
+        properties?: Property[]
     }
 }
